Move Files to the right Place
