--- Silicon/QcomPkg/Library/ArmPlatformLib/InitializationUtils.h
@@ -1,6 +1,0 @@
-#ifndef _INITIALIZATION_UTILS_H_
-#define _INITIALIZATION_UTILS_H_
-
-VOID EarlyInitialization(VOID);
-
-#endif /* _INITIALIZATION_UTILS_H_ */
