+++ Silicon/Qualcomm/QcomPkg/Library/ArmPlatformLib/InitializationUtils.c
@@ -42,11 +42,6 @@ VOID EarlyInitialization(VOID)
   // Initialize UART Serial
   UartInit();
 
-  // Initialize GIC
-  MmioWrite32(
-      GICR_WAKER_CURRENT_CPU,
-      (MmioRead32(GICR_WAKER_CURRENT_CPU) & ~GIC_WAKER_PROCESSORSLEEP));
-
   // Do platform specific initialization here
   PlatformInitialize();
 }
