+++ Silicon/Qualcomm/SM6115Pkg/Library/PlatformPrePiLib/PlatformUtils.c
@@ -3,4 +3,13 @@
 #include <Library/PlatformPrePiLib.h>
 #include "PlatformUtils.h"
 
-VOID PlatformInitialize(VOID) {}
+VOID SetWatchdogState(BOOLEAN Enable)
+{
+  MmioWrite32(APSS_WDT_BASE + APSS_WDT_ENABLE_OFFSET, Enable);
+}
+
+VOID PlatformInitialize(VOID)
+{
+  // Disable WatchDog Timer
+  SetWatchdogState(FALSE);
+}
