+++ Silicon/Qualcomm/SM8450Pkg/Library/PlatformPrePiLib/PlatformUtils.c
@@ -3,4 +3,10 @@
 #include <Library/PlatformPrePiLib.h>
 #include "PlatformUtils.h"
 
-VOID PlatformInitialize(VOID) {}
+VOID PlatformInitialize(VOID)
+{
+  // Initialize GIC
+  MmioWrite32(
+      GICR_WAKER_CURRENT_CPU,
+      (MmioRead32(GICR_WAKER_CURRENT_CPU) & ~GIC_WAKER_PROCESSORSLEEP));
+}
