--- Silicon/SM7150Pkg/Library/PlatformPrePiLib/PlatformUtils.h
@@ -1,7 +1,0 @@
-#ifndef _PLATFORM_UTILS_H_
-#define _PLATFORM_UTILS_H_
-
-#define APSS_WDT_BASE 0x17c10000
-#define APSS_WDT_ENABLE_OFFSET 0x8
-
-#endif /* _PLATFORM_UTILS_H_ */
