--- Silicon/SM8350Pkg/Library/PlatformPrePiLib/PlatformUtils.c
@@ -1,12 +1,0 @@
-#include <PiPei.h>
-#include <Library/IoLib.h>
-#include <Library/PlatformPrePiLib.h>
-#include "PlatformUtils.h"
-
-VOID PlatformInitialize(VOID)
-{
-  // Initialize GIC
-  MmioWrite32(
-      GICR_WAKER_CURRENT_CPU,
-      (MmioRead32(GICR_WAKER_CURRENT_CPU) & ~GIC_WAKER_PROCESSORSLEEP));
-}
